Change slightly the way integers are printed by printf() and friends.
Previously, %d printed a 32-bit int.  %ld and %lld printed 64-bit ints.
So if you wanted to print a word-sized int (eg. a SizeT variable), you
had to cast it to a Long and then print with %lld in order to work on
both 32-bit and 64-bit platforms.

I changed things so that %d prints a 32-bit int, %ld prints a word-sized
int, and %lld prints a 64-bit int.  There are two advantages to this:
- it now matches the way the normal glibc printf() works;
- you can print word-sized ints without casting.

I also made the corresponding change for %u/lu/llu and %x/lx/llx, and I
changed a couple of VG_(printf)() invocations accordingly.





git-svn-id: svn://svn.valgrind.org/valgrind/trunk@4527

+++ coregrind/m_aspacemgr/aspacemgr.c
@@ -741,9 +741,9 @@ VG_(map_file_segment)( Addr addr, SizeT len,
       VG_(printf)(
          "\n"
          "map_file_segment(addr=%p len=%lu prot=0x%x flags=0x%x\n"
-         "                 dev=0x%4x ino=%d off=%ld\n"
+         "                 dev=0x%4x ino=%d off=%lld\n"
          "                 filename='%s')\n",
-         addr, (ULong)len, prot, flags, dev, ino, off, filename);
+         addr, len, prot, flags, dev, ino, off, filename);
 
    if (0) VG_(show_segments)("before map_file_segment");
